#33: test refactoring
