+++ database/src/test/resources/custom-config.yaml
@@ -1,0 +1,18 @@
+port: 8800
+nodeId: 11
+logDir: /var/logs2
+cluster: # comment
+  - node1
+  - node2
+  - node3
+  - node4
+  - node5
+  - node6
+  - node7
+  - node8
+  - node9
+  - node10
+  - node11
+
+unusedProperty:
+  enabled: true
